# Use all modified files when calculating changed targets