--- a/package.json
+++ b/package.json
@@ -1,6 +1,6 @@
 {
   "name": "jazelle",
-  "version": "0.0.0-standalone.93",
+  "version": "0.0.0-standalone.94",
   "main": "index.js",
   "bin": {
     "barn": "bin/bootstrap.sh",