--- a/utils/find-changed-targets.js
+++ b/utils/find-changed-targets.js
@@ -43,7 +43,7 @@ const scan = async (root, lines) => {
 };
 
 function quoteFilePaths(paths) {
-  return paths.map(path => `'${path}'`);
+  return paths.map(path => `"${path}"`);
 }
 
 const findChangedBazelTargets = async ({root, files}) => {
@@ -64,8 +64,17 @@ const findChangedBazelTargets = async ({root, files}) => {
     .map(line => line.trim())
     .filter(line => !ignored.find(i => line.startsWith(i)));
 
-  const invalid = lines.find(line => line.includes(' '));
-  if (invalid) throw new Error(`File path cannot contain spaces: ${invalid}`);
+  const invalid = lines.find(line => line.includes('"'));
+  if (invalid) {
+    // Disallow double quote (") character in file paths, because Bazel query syntax
+    // utilizes double quotes to encapsulate strings. Including a double quote within
+    // a file path could lead to syntax errors in the Bazel query.
+    // Note: In the future, we may relax this constraint if support is required. This could
+    // be achieved by running multiple queries, where file paths containing special characters
+    // like single or double quotes could be queried separately. File paths with single quotes
+    // could be encapsulated using double quotes, and vice versa, to ensure query integrity.
+    throw new Error(`File path cannot contain double quotes: '${invalid}'`);
+  }
 
   const {projects, workspace} = await getManifest({root});
   if (workspace === 'sandbox') {
@@ -105,7 +114,7 @@ const findChangedBazelTargets = async ({root, files}) => {
               const regex = /not declared in package '(.*?)'/g;
               return Array.from(e.message.matchAll(regex))
                 .map(([, pkg]) =>
-                  pkg ? `kind("source file", //${pkg}:*)` : ''
+                  pkg ? `kind("source file", "//${pkg}:*")` : ''
                 )
                 .filter(Boolean);
             })