Remove an optimization logic from the findChangedTargets function. This optimization was originally designed to reduce query size by considering only one JS/TS file per directory, based on the assumption that all source files in a directory would map to the same target. However, this assumption does not hold when working with more granular targets, necessitating the retention of all files to calculate changed targets more accurately.

By removing this optimization, we ensure correct file handling for all types of files and targets. The query size limitation is no longer a concern since we switched to using query files. Despite the removal, there is no significant speed regression observed in our monorepo containing hundreds of thousands of files. Essentially, we are trading speed for correctness.

Also, wrap file paths in quotes to handle special characters in bazel query, disallowing double quotes (") in file paths (we may relax this constraint in the future if support is required)

```
% time jz bazel query 'kind("source file", //...:*)' --output=label | sed 's|^//||; s|:|/|' | jz changes > /dev/null
# Before
jz changes 3.56s user 2.42s system 48% cpu 12.289 total

# After
jz changes 6.10s user 4.67s system 65% cpu 16.531 total
```